// frontend/src/api/generated/models/user-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UserDetail } from './user-detail';

/**
 * 
 * @export
 * @interface UserPageResponse
 */
export interface UserPageResponse {
    /**
     * 
     * @type {Array<UserDetail>}
     * @memberof UserPageResponse
     */
    'content': Array<UserDetail>;
    /**
     * 
     * @type {number}
     * @memberof UserPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof UserPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof UserPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof UserPageResponse
     */
    'totalPages': number;
}